server/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-meal-planner"
version = "0.1.0"
description = "AI-powered meal planner application"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "sqlalchemy",
    "alembic",
    "uvicorn",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "python-multipart",
    "requests",
]

[tool.hatch.build.targets.wheel]
packages = ["app"] 